+++ api/src/modules/comments-area/comments-area-request/dtos/comments-area-request-rejected-in.dto.ts
@@ -1,0 +1,7 @@
+import { IsOptional, IsString } from 'class-validator';
+
+export class CommentsAreaRequestRejectedInDto {
+  @IsString()
+  @IsOptional()
+  reason?: string;
+}