+++ api/src/modules/notification/notification-type.ts
@@ -1,4 +1,6 @@
 export enum NotificationType {
   RULES_UPDATE = 'rulesUpdate',
   SUBSCRIPTION_REPLY = 'subscriptionReply',
+  COMMENTS_AREA_REQUEST_APPROVED = 'commentsAreaRequestApproved',
+  COMMENTS_AREA_REQUEST_REJECTED = 'commentsAreaRequestRejected',
 }